Extract InspectorTabs from the Inspector mega-component
Begin splitting the 1189-line Inspector by moving the self-contained tab strip
(InspectorTab type, TAB_DEFS, InspectorTabs presentational component) into
InspectorTabs.tsx and importing it. Behavior-preserving; typecheck clean.
Footer panels (Fal/Kie) and the I/O section can follow the same pattern.
