+++ web/src/components/Inspector.tsx
@@ -37,6 +37,7 @@ import { useExposedInputToggle } from "../hooks/nodes/useExposedInputToggle";
 import usePropertyValidationStore from "../stores/PropertyValidationStore";
 import { useStoreWithEqualityFn } from "zustand/traditional";
 import RunSelectedNodesSection from "./inspector/RunSelectedNodesSection";
+import { InspectorTabs, type InspectorTab } from "./InspectorTabs";
 import { colorForType } from "../config/data_types";
 import { IconForType } from "../config/IconForType";
 
@@ -501,48 +502,6 @@ const ValidationErrorBanner: React.FC<ValidationErrorBannerProps> = ({
     </div>
   );
 };
-
-type InspectorTab = "params" | "io" | "help";
-
-const TAB_DEFS: { value: InspectorTab; label: string; hasCount: boolean }[] = [
-  { value: "params", label: "Params", hasCount: true },
-  { value: "io", label: "I/O", hasCount: true },
-  { value: "help", label: "Help", hasCount: false }
-];
-
-interface InspectorTabsProps {
-  active: InspectorTab;
-  onChange: (next: InspectorTab) => void;
-  counts: Partial<Record<InspectorTab, number>>;
-}
-
-const InspectorTabs: React.FC<InspectorTabsProps> = ({
-  active,
-  onChange,
-  counts
-}) => (
-  <div className="inspector-tabs" role="tablist">
-    {TAB_DEFS.map((tab) => {
-      const count = counts[tab.value];
-      const isActive = active === tab.value;
-      return (
-        <button
-          key={tab.value}
-          type="button"
-          role="tab"
-          aria-selected={isActive}
-          className={`inspector-tab${isActive ? " is-active" : ""}`}
-          onClick={() => onChange(tab.value)}
-        >
-          {tab.label}
-          {tab.hasCount && typeof count === "number" ? (
-            <span className="tab-count">{count}</span>
-          ) : null}
-        </button>
-      );
-    })}
-  </div>
-);
 
 const TypeLabel: React.FC<{ type?: TypeMetadata | null }> = ({ type }) => {
   const label = type?.type ?? "any";
